pyproject.toml
[project]
name = "kindle2keynote"
version = "0.1.0"
description = "Convert PDF ebooks to Marp presentation slides using AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyMuPDF>=1.24.0",
    "pdfplumber>=0.11.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
kindle2keynote = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kindle2keynote"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
